Stack2: Complete
I could not get only the command Env to work:
cmdTarget.Env = append(cmdTarget.Env, "environmentKey"+"="+string(password))

So instead I set the entire environment.

And I need to get the StdErr out to get some of the messages I believe.
At the moment my deployment script just ignores all of it, and some of
the messages when you fail to get the key is sent as error.

To see the source code check:
https://exploit.education/protostar/stack-two/