Migrate to Phoenix
As it turned out Phoenix is a much more pleasant experience.

The OS has history (Ctrl+R, or even arrow_up), and overall is
better in probably a lot of ways.

The new source code pages are here:
https://exploit.education/phoenix/

I had to tweak the solutions a bit, but the result is overall
the same.